# src/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "0.1.0",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "elinge-boneless.com/api",
    "basePath": "",
...
    "schemes": [
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/clientes/eliminar/{idCliente}": {
            "delete": {
                "summary": "Eliminar un cliente en específico de la página El Inge Boneless.",
                "description": "<p>Servicio que elimina el cliente dentro de la página El Inge Boneless.</p>",
                "operationId": "eliminar",
                "parameters": [
                    {
                        "name": "idCliente",
                        "in": "path",
                        "description": "id del Cliente a eliminar",
                        "required": true,
                        "default": 5
                    }
                ],
                "tags": [
                    "Clientes"
                ],
                "responses": {
                    "200": {
                        "description": "Cliente Eliminado."
                    },
                    "201": {
                        "description": "respuesta vacía"
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                }
            }
        },
        "/contenidoPedido/eliminar/{idOrden}": {
            "delete": {
                "summary": "Elimina el contenido de un pedido en específico de la página El Inge Boneless",
                "description": "<p>Servicio que elimina el contenido de un pedido dentro de la página El Inge Boneless.</p>",
                "operationId": "eliminar",
                "parameters": [
                    {
                        "name": "idOrden",
                        "in": "path",
                        "descrition": "id de la orden a eliminar",
                        "required": true,
                        "default": "8"
                    }
                ],
                "tags": [
                    "Contenido Pedido"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "201": {
                        "description": "Respuesta vacía"
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                }
            }
        },
        "/clientes/actualizar/{idCliente}": {
            "put": {
                "summary": "Modificando un cliente en específico de la página El Inge Boneless.",
                "description": "<p>Servicio que modifica el cliente de la página El Inge Boneless.</p>",
                "operationId": "actualizar",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "idCliente",
                        "in": "path",
                        "description": "id del Cliente a actualizar",
                        "required": true,
                        "default": 1
                    },
                    {
                        "name": "Cliente",
                        "in": "body",
                        "description": "Cliente a modificar",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Cliente desactualizado"
                        }
                    }
                ],
                "tags": [
                    "Clientes"
                ],
                "responses": {
                    "200": {
                        "description": "Cliente Modificado."
                    },
                    "201": {
                        "description": "respuesta vacía"
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                }
            }
        },
        "/contenidoPedido/actualizar/{idOrden}": {
            "put": {
                "summary": "Actualiza el contenido de un pedido en específico de la página El Inge Boneless",
                "description": "<p>Servicio que actualiza el contenido de un determinado pedido dentro de la página El Inge Boneless.</p>",
                "operationId": "actualizar",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "idOrden",
                        "in": "path",
                        "description": "id de la orden a cambiar",
                        "required": true,
                        "default": 11
                    },
                    {
                        "name": "Contenido pedido",
                        "in": "body",
                        "description": "Contenido pedido a modificar",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Contenido pedido desactualizado"
                        }
                    }
                ],
                "tags": [
                    "Contenido Pedido"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "201": {
                        "description": "Respuesta vacía"
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                }
            }
        },
...
        "/pedidos/eliminar/{idPedido}": {
            "delete": {
                "summary": "Elimina el pedido en específico de la página El Inge Boneless",
                "description": "<p>Servicio que elimina el pedido dentro de la página El Inge Boneless.</p>",
                "operationId": "eliminar",
                "parameters": [
                    {
                        "name": "idPedido",
                        "in": "path",
                        "descrition": "id del pedido a eliminar",
                        "required": true,
                        "default": "8"
                    }
                ],
                "tags": [
                    "Pedido"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "201": {
                        "description": "Respuesta vacía"
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                }
            }
        },
        "/pedidos/actualizar/{idPedido}": {
            "put": {
                "summary": "Actualiza el pedido en específico de la página El Inge Boneless",
                "description": "<p>Servicio que actualiza un determinado pedido dentro de la página El Inge Boneless.</p>",
                "operationId": "actualizar",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "idPedido",
                        "in": "path",
                        "description": "id del pedido a cambiar",
                        "required": true,
                        "default": 11
                    },
                    {
                        "name": "Pedido",
                        "in": "body",
                        "description": "Pedido a modificar",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Pedido desactualizado"
                        }
                    }
                ],
                "tags": [
                    "Pedido"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "201": {
                        "description": "Respuesta vacía"
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Cliente": {
            "type": "object",
            "properties": {
                "idCliente": {
                    "type": "number",
                    "description": "Identificador del cliente"
                },
                "nombre": {
                    "type": "string",
                    "description": "Nombre del cliente"
                },
                "apellidos": {
                    "type": "string",
                    "description": "Apellidos del cliente"
                },
                "telefono": {
                    "type": "string",
                    "description": "Telefono del cliente"
                },
                "correo": {
                    "type": "string",
                    "description": "Correo del cliente"
                },
                "contraseña": {
                    "type": "string",
                    "description": "Contraseña del cliente"
                },
                "sexo": {
                    "type": "string",
                    "description": "Sexo del cliente"
                }
            }
        },
...
        "Pedido": {
            "type": "object",
            "properties": {
                "idPedido": {
                    "type": "number",
                    "description": "Identificador del número de pedido"
                },
                "idCliente": {
                    "type": "number",
                    "description": "Identificador del cliente"
                },
                "telefonoExtra": {
                    "type": "text",
                    "description": "Telefono extra para el pedido"
                },
                "total": {
                    "type": "number",
                    "description": "Total de dinero por el pedido"
                },
                "dineroRecibido": {
                    "type": "number",
                    "description": "Dinero recibido por parte del cliente"
                },
                "notas": {
                    "type": "string",
                    "description": "Notas para realizar el pedido"
                },
                "fecha": {
                    "type": "string",
                    "description": "Fecha a realizar el pedido"
                },
                "finalizado": {
                    "type": "number",
                    "description": "Estado del pedido 0 si se encuentra en proceso, 1 si ya fue realizado"
                }
            }
        },
...
    }
}